Update Products.jsx
The Products component is responsible for displaying a list of products. It receives an array of product data from the App.js component and maps through it to render individual ProductItem components.

Responsibilities:
Accepts a list of products as props.
Maps over the product data and renders multiple ProductItem components.
Acts as a container for displaying all the products.